Move std::io to alloc
