+++ library/alloc/src/io/buffered/linewriter.rs
