+++ library/alloc/src/io/buffered/tests.rs
