+++ library/std/src/io/mod.rs
@@ -302,18 +302,21 @@ pub use alloc::io::RawOsError;
 #[doc(hidden)]
 #[unstable(feature = "io_const_error_internals", issue = "none")]
 pub use alloc::io::SimpleMessage;
+#[stable(feature = "bufwriter_into_parts", since = "1.56.0")]
+pub use alloc::io::WriterPanicked;
 #[unstable(feature = "io_const_error", issue = "133448")]
 pub use alloc::io::const_error;
 #[stable(feature = "rust1", since = "1.0.0")]
-pub use alloc::io::{Error, ErrorKind, Result};
+pub use alloc::io::{
+    BufRead, BufReader, BufWriter, Bytes, Chain, Cursor, Empty, Error, ErrorKind, IntoInnerError,
+    LineWriter, Lines, Read, Repeat, Result, Seek, SeekFrom, Sink, Split, Take, Write, empty,
+    prelude, read_to_string, repeat, sink,
+};
 #[stable(feature = "iovec", since = "1.36.0")]
 pub use alloc::io::{IoSlice, IoSliceMut};
 #[unstable(feature = "read_buf", issue = "78485")]
 pub use core::io::{BorrowedBuf, BorrowedCursor};
-use core::slice::memchr;
 
-#[stable(feature = "bufwriter_into_parts", since = "1.56.0")]
-pub use self::buffered::WriterPanicked;
 #[stable(feature = "anonymous_pipe", since = "1.87.0")]
 pub use self::pipe::{PipeReader, PipeWriter, pipe};
 #[stable(feature = "is_terminal", since = "1.70.0")]
@@ -327,2768 +330,18 @@ pub use self::stdio::{_eprint, _print};
 pub use self::stdio::{set_output_capture, try_set_output_capture};
 #[stable(feature = "rust1", since = "1.0.0")]
 pub use self::{
-    buffered::{BufReader, BufWriter, IntoInnerError, LineWriter},
     copy::copy,
-    cursor::Cursor,
     stdio::{Stderr, StderrLock, Stdin, StdinLock, Stdout, StdoutLock, stderr, stdin, stdout},
-    util::{Empty, Repeat, Sink, empty, repeat, sink},
 };
-use crate::mem::MaybeUninit;
-use crate::{cmp, fmt, slice, str};
 
-mod buffered;
 pub(crate) mod copy;
-mod cursor;
 mod error;
-mod impls;
 mod pipe;
-pub mod prelude;
 mod stdio;
-mod util;
 
-const DEFAULT_BUF_SIZE: usize = crate::sys::io::DEFAULT_BUF_SIZE;
+pub(crate) use alloc::io::{
+    SpecReadByte, default_read_buf, default_read_to_end, default_read_to_string,
+    default_read_vectored, default_write_vectored, stream_len_default,
+};
 
 pub(crate) use stdio::cleanup;
-
-struct Guard<'a> {
-    buf: &'a mut Vec<u8>,
-    len: usize,
-}
-
-impl Drop for Guard<'_> {
-    fn drop(&mut self) {
-        unsafe {
-            self.buf.set_len(self.len);
-        }
-    }
-}
-
-// Several `read_to_string` and `read_line` methods in the standard library will
-// append data into a `String` buffer, but we need to be pretty careful when
-// doing this. The implementation will just call `.as_mut_vec()` and then
-// delegate to a byte-oriented reading method, but we must ensure that when
-// returning we never leave `buf` in a state such that it contains invalid UTF-8
-// in its bounds.
-//
-// To this end, we use an RAII guard (to protect against panics) which updates
-// the length of the string when it is dropped. This guard initially truncates
-// the string to the prior length and only after we've validated that the
-// new contents are valid UTF-8 do we allow it to set a longer length.
-//
-// The unsafety in this function is twofold:
-//
-// 1. We're looking at the raw bytes of `buf`, so we take on the burden of UTF-8
-//    checks.
-// 2. We're passing a raw buffer to the function `f`, and it is expected that
-//    the function only *appends* bytes to the buffer. We'll get undefined
-//    behavior if existing bytes are overwritten to have non-UTF-8 data.
-pub(crate) unsafe fn append_to_string<F>(buf: &mut String, f: F) -> Result<usize>
-where
-    F: FnOnce(&mut Vec<u8>) -> Result<usize>,
-{
-    let mut g = Guard { len: buf.len(), buf: unsafe { buf.as_mut_vec() } };
-    let ret = f(g.buf);
-
-    // SAFETY: the caller promises to only append data to `buf`
-    let appended = unsafe { g.buf.get_unchecked(g.len..) };
-    if str::from_utf8(appended).is_err() {
-        ret.and_then(|_| Err(Error::INVALID_UTF8))
-    } else {
-        g.len = g.buf.len();
-        ret
-    }
-}
-
-// Here we must serve many masters with conflicting goals:
-//
-// - avoid allocating unless necessary
-// - avoid overallocating if we know the exact size (#89165)
-// - avoid passing large buffers to readers that always initialize the free capacity if they perform short reads (#23815, #23820)
-// - pass large buffers to readers that do not initialize the spare capacity. this can amortize per-call overheads
-// - and finally pass not-too-small and not-too-large buffers to Windows read APIs because they manage to suffer from both problems
-//   at the same time, i.e. small reads suffer from syscall overhead, all reads incur costs proportional to buffer size (#110650)
-//
-pub(crate) fn default_read_to_end<R: Read + ?Sized>(
-    r: &mut R,
-    buf: &mut Vec<u8>,
-    size_hint: Option<usize>,
-) -> Result<usize> {
-    let start_len = buf.len();
-    let start_cap = buf.capacity();
-    // Optionally limit the maximum bytes read on each iteration.
-    // This adds an arbitrary fiddle factor to allow for more data than we expect.
-    let mut max_read_size = size_hint
-        .and_then(|s| s.checked_add(1024)?.checked_next_multiple_of(DEFAULT_BUF_SIZE))
-        .unwrap_or(DEFAULT_BUF_SIZE);
-
-    let mut initialized = 0; // Extra initialized bytes from previous loop iteration
-
-    const PROBE_SIZE: usize = 32;
-
-    fn small_probe_read<R: Read + ?Sized>(r: &mut R, buf: &mut Vec<u8>) -> Result<usize> {
-        let mut probe = [0u8; PROBE_SIZE];
-
-        loop {
-            match r.read(&mut probe) {
-                Ok(n) => {
-                    // there is no way to recover from allocation failure here
-                    // because the data has already been read.
-                    buf.extend_from_slice(&probe[..n]);
-                    return Ok(n);
-                }
-                Err(ref e) if e.is_interrupted() => continue,
-                Err(e) => return Err(e),
-            }
-        }
-    }
-
-    // avoid inflating empty/small vecs before we have determined that there's anything to read
-    if (size_hint.is_none() || size_hint == Some(0)) && buf.capacity() - buf.len() < PROBE_SIZE {
-        let read = small_probe_read(r, buf)?;
-
-        if read == 0 {
-            return Ok(0);
-        }
-    }
-
-    let mut consecutive_short_reads = 0;
-
-    loop {
-        if buf.len() == buf.capacity() && buf.capacity() == start_cap {
-            // The buffer might be an exact fit. Let's read into a probe buffer
-            // and see if it returns `Ok(0)`. If so, we've avoided an
-            // unnecessary doubling of the capacity. But if not, append the
-            // probe buffer to the primary buffer and let its capacity grow.
-            let read = small_probe_read(r, buf)?;
-
-            if read == 0 {
-                return Ok(buf.len() - start_len);
-            }
-        }
-
-        if buf.len() == buf.capacity() {
-            // buf is full, need more space
-            buf.try_reserve(PROBE_SIZE)?;
-        }
-
-        let mut spare = buf.spare_capacity_mut();
-        let buf_len = cmp::min(spare.len(), max_read_size);
-        spare = &mut spare[..buf_len];
-        let mut read_buf: BorrowedBuf<'_> = spare.into();
-
-        // SAFETY: These bytes were initialized but not filled in the previous loop
-        unsafe {
-            read_buf.set_init(initialized);
-        }
-
-        let mut cursor = read_buf.unfilled();
-        let result = loop {
-            match r.read_buf(cursor.reborrow()) {
-                Err(e) if e.is_interrupted() => continue,
-                // Do not stop now in case of error: we might have received both data
-                // and an error
-                res => break res,
-            }
-        };
-
-        let unfilled_but_initialized = cursor.init_mut().len();
-        let bytes_read = cursor.written();
-        let was_fully_initialized = read_buf.init_len() == buf_len;
-
-        // SAFETY: BorrowedBuf's invariants mean this much memory is initialized.
-        unsafe {
-            let new_len = bytes_read + buf.len();
-            buf.set_len(new_len);
-        }
-
-        // Now that all data is pushed to the vector, we can fail without data loss
-        result?;
-
-        if bytes_read == 0 {
-            return Ok(buf.len() - start_len);
-        }
-
-        if bytes_read < buf_len {
-            consecutive_short_reads += 1;
-        } else {
-            consecutive_short_reads = 0;
-        }
-
-        // store how much was initialized but not filled
-        initialized = unfilled_but_initialized;
-
-        // Use heuristics to determine the max read size if no initial size hint was provided
-        if size_hint.is_none() {
-            // The reader is returning short reads but it doesn't call ensure_init().
-            // In that case we no longer need to restrict read sizes to avoid
-            // initialization costs.
-            // When reading from disk we usually don't get any short reads except at EOF.
-            // So we wait for at least 2 short reads before uncapping the read buffer;
-            // this helps with the Windows issue.
-            if !was_fully_initialized && consecutive_short_reads > 1 {
-                max_read_size = usize::MAX;
-            }
-
-            // we have passed a larger buffer than previously and the
-            // reader still hasn't returned a short read
-            if buf_len >= max_read_size && bytes_read == buf_len {
-                max_read_size = max_read_size.saturating_mul(2);
-            }
-        }
-    }
-}
-
-pub(crate) fn default_read_to_string<R: Read + ?Sized>(
-    r: &mut R,
-    buf: &mut String,
-    size_hint: Option<usize>,
-) -> Result<usize> {
-    // Note that we do *not* call `r.read_to_end()` here. We are passing
-    // `&mut Vec<u8>` (the raw contents of `buf`) into the `read_to_end`
-    // method to fill it up. An arbitrary implementation could overwrite the
-    // entire contents of the vector, not just append to it (which is what
-    // we are expecting).
-    //
-    // To prevent extraneously checking the UTF-8-ness of the entire buffer
-    // we pass it to our hardcoded `default_read_to_end` implementation which
-    // we know is guaranteed to only read data into the end of the buffer.
-    unsafe { append_to_string(buf, |b| default_read_to_end(r, b, size_hint)) }
-}
-
-pub(crate) fn default_read_vectored<F>(read: F, bufs: &mut [IoSliceMut<'_>]) -> Result<usize>
-where
-    F: FnOnce(&mut [u8]) -> Result<usize>,
-{
-    let buf = bufs.iter_mut().find(|b| !b.is_empty()).map_or(&mut [][..], |b| &mut **b);
-    read(buf)
-}
-
-pub(crate) fn default_write_vectored<F>(write: F, bufs: &[IoSlice<'_>]) -> Result<usize>
-where
-    F: FnOnce(&[u8]) -> Result<usize>,
-{
-    let buf = bufs.iter().find(|b| !b.is_empty()).map_or(&[][..], |b| &**b);
-    write(buf)
-}
-
-pub(crate) fn default_read_exact<R: Read + ?Sized>(this: &mut R, mut buf: &mut [u8]) -> Result<()> {
-    while !buf.is_empty() {
-        match this.read(buf) {
-            Ok(0) => break,
-            Ok(n) => {
-                buf = &mut buf[n..];
-            }
-            Err(ref e) if e.is_interrupted() => {}
-            Err(e) => return Err(e),
-        }
-    }
-    if !buf.is_empty() { Err(Error::READ_EXACT_EOF) } else { Ok(()) }
-}
-
-pub(crate) fn default_read_buf<F>(read: F, mut cursor: BorrowedCursor<'_>) -> Result<()>
-where
-    F: FnOnce(&mut [u8]) -> Result<usize>,
-{
-    let n = read(cursor.ensure_init().init_mut())?;
-    cursor.advance(n);
-    Ok(())
-}
-
-pub(crate) fn default_read_buf_exact<R: Read + ?Sized>(
-    this: &mut R,
-    mut cursor: BorrowedCursor<'_>,
-) -> Result<()> {
-    while cursor.capacity() > 0 {
-        let prev_written = cursor.written();
-        match this.read_buf(cursor.reborrow()) {
-            Ok(()) => {}
-            Err(e) if e.is_interrupted() => continue,
-            Err(e) => return Err(e),
-        }
-
-        if cursor.written() == prev_written {
-            return Err(Error::READ_EXACT_EOF);
-        }
-    }
-
-    Ok(())
-}
-
-pub(crate) fn default_write_fmt<W: Write + ?Sized>(
-    this: &mut W,
-    args: fmt::Arguments<'_>,
-) -> Result<()> {
-    // Create a shim which translates a `Write` to a `fmt::Write` and saves off
-    // I/O errors, instead of discarding them.
-    struct Adapter<'a, T: ?Sized + 'a> {
-        inner: &'a mut T,
-        error: Result<()>,
-    }
-
-    impl<T: Write + ?Sized> fmt::Write for Adapter<'_, T> {
-        fn write_str(&mut self, s: &str) -> fmt::Result {
-            match self.inner.write_all(s.as_bytes()) {
-                Ok(()) => Ok(()),
-                Err(e) => {
-                    self.error = Err(e);
-                    Err(fmt::Error)
-                }
-            }
-        }
-    }
-
-    let mut output = Adapter { inner: this, error: Ok(()) };
-    match fmt::write(&mut output, args) {
-        Ok(()) => Ok(()),
-        Err(..) => {
-            // Check whether the error came from the underlying `Write`.
-            if output.error.is_err() {
-                output.error
-            } else {
-                // This shouldn't happen: the underlying stream did not error,
-                // but somehow the formatter still errored?
-                panic!(
-                    "a formatting trait implementation returned an error when the underlying stream did not"
-                );
-            }
-        }
-    }
-}
-
-/// The `Read` trait allows for reading bytes from a source.
-///
-/// Implementors of the `Read` trait are called 'readers'.
-///
-/// Readers are defined by one required method, [`read()`]. Each call to [`read()`]
-/// will attempt to pull bytes from this source into a provided buffer. A
-/// number of other methods are implemented in terms of [`read()`], giving
-/// implementors a number of ways to read bytes while only needing to implement
-/// a single method.
-///
-/// Readers are intended to be composable with one another. Many implementors
-/// throughout [`std::io`] take and provide types which implement the `Read`
-/// trait.
-///
-/// Please note that each call to [`read()`] may involve a system call, and
-/// therefore, using something that implements [`BufRead`], such as
-/// [`BufReader`], will be more efficient.
-///
-/// Repeated calls to the reader use the same cursor, so for example
-/// calling `read_to_end` twice on a [`File`] will only return the file's
-/// contents once. It's recommended to first call `rewind()` in that case.
-///
-/// # Examples
-///
-/// [`File`]s implement `Read`:
-///
-/// ```no_run
-/// use std::io;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> io::Result<()> {
-///     let mut f = File::open("foo.txt")?;
-///     let mut buffer = [0; 10];
-///
-///     // read up to 10 bytes
-///     f.read(&mut buffer)?;
-///
-///     let mut buffer = Vec::new();
-///     // read the whole file
-///     f.read_to_end(&mut buffer)?;
-///
-///     // read into a String, so that you don't need to do the conversion.
-///     let mut buffer = String::new();
-///     f.read_to_string(&mut buffer)?;
-///
-///     // and more! See the other methods for more details.
-///     Ok(())
-/// }
-/// ```
-///
-/// Read from [`&str`] because [`&[u8]`][prim@slice] implements `Read`:
-///
-/// ```no_run
-/// # use std::io;
-/// use std::io::prelude::*;
-///
-/// fn main() -> io::Result<()> {
-///     let mut b = "This string will be read".as_bytes();
-///     let mut buffer = [0; 10];
-///
-///     // read up to 10 bytes
-///     b.read(&mut buffer)?;
-///
-///     // etc... it works exactly as a File does!
-///     Ok(())
-/// }
-/// ```
-///
-/// [`read()`]: Read::read
-/// [`&str`]: prim@str
-/// [`std::io`]: self
-/// [`File`]: crate::fs::File
-#[stable(feature = "rust1", since = "1.0.0")]
-#[doc(notable_trait)]
-#[cfg_attr(not(test), rustc_diagnostic_item = "IoRead")]
-pub trait Read {
-    /// Pull some bytes from this source into the specified buffer, returning
-    /// how many bytes were read.
-    ///
-    /// This function does not provide any guarantees about whether it blocks
-    /// waiting for data, but if an object needs to block for a read and cannot,
-    /// it will typically signal this via an [`Err`] return value.
-    ///
-    /// If the return value of this method is [`Ok(n)`], then implementations must
-    /// guarantee that `0 <= n <= buf.len()`. A nonzero `n` value indicates
-    /// that the buffer `buf` has been filled in with `n` bytes of data from this
-    /// source. If `n` is `0`, then it can indicate one of two scenarios:
-    ///
-    /// 1. This reader has reached its "end of file" and will likely no longer
-    ///    be able to produce bytes. Note that this does not mean that the
-    ///    reader will *always* no longer be able to produce bytes. As an example,
-    ///    on Linux, this method will call the `recv` syscall for a [`TcpStream`],
-    ///    where returning zero indicates the connection was shut down correctly. While
-    ///    for [`File`], it is possible to reach the end of file and get zero as result,
-    ///    but if more data is appended to the file, future calls to `read` will return
-    ///    more data.
-    /// 2. The buffer specified was 0 bytes in length.
-    ///
-    /// It is not an error if the returned value `n` is smaller than the buffer size,
-    /// even when the reader is not at the end of the stream yet.
-    /// This may happen for example because fewer bytes are actually available right now
-    /// (e. g. being close to end-of-file) or because read() was interrupted by a signal.
-    ///
-    /// As this trait is safe to implement, callers in unsafe code cannot rely on
-    /// `n <= buf.len()` for safety.
-    /// Extra care needs to be taken when `unsafe` functions are used to access the read bytes.
-    /// Callers have to ensure that no unchecked out-of-bounds accesses are possible even if
-    /// `n > buf.len()`.
-    ///
-    /// *Implementations* of this method can make no assumptions about the contents of `buf` when
-    /// this function is called. It is recommended that implementations only write data to `buf`
-    /// instead of reading its contents.
-    ///
-    /// Correspondingly, however, *callers* of this method in unsafe code must not assume
-    /// any guarantees about how the implementation uses `buf`. The trait is safe to implement,
-    /// so it is possible that the code that's supposed to write to the buffer might also read
-    /// from it. It is your responsibility to make sure that `buf` is initialized
-    /// before calling `read`. Calling `read` with an uninitialized `buf` (of the kind one
-    /// obtains via [`MaybeUninit<T>`]) is not safe, and can lead to undefined behavior.
-    ///
-    /// [`MaybeUninit<T>`]: crate::mem::MaybeUninit
-    ///
-    /// # Errors
-    ///
-    /// If this function encounters any form of I/O or other error, an error
-    /// variant will be returned. If an error is returned then it must be
-    /// guaranteed that no bytes were read.
-    ///
-    /// An error of the [`ErrorKind::Interrupted`] kind is non-fatal and the read
-    /// operation should be retried if there is nothing else to do.
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`Ok(n)`]: Ok
-    /// [`File`]: crate::fs::File
-    /// [`TcpStream`]: crate::net::TcpStream
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///     let mut buffer = [0; 10];
-    ///
-    ///     // read up to 10 bytes
-    ///     let n = f.read(&mut buffer[..])?;
-    ///
-    ///     println!("The bytes: {:?}", &buffer[..n]);
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize>;
-
-    /// Like `read`, except that it reads into a slice of buffers.
-    ///
-    /// Data is copied to fill each buffer in order, with the final buffer
-    /// written to possibly being only partially filled. This method must
-    /// behave equivalently to a single call to `read` with concatenated
-    /// buffers.
-    ///
-    /// The default implementation calls `read` with either the first nonempty
-    /// buffer provided, or an empty one if none exists.
-    #[stable(feature = "iovec", since = "1.36.0")]
-    fn read_vectored(&mut self, bufs: &mut [IoSliceMut<'_>]) -> Result<usize> {
-        default_read_vectored(|b| self.read(b), bufs)
-    }
-
-    /// Determines if this `Read`er has an efficient `read_vectored`
-    /// implementation.
-    ///
-    /// If a `Read`er does not override the default `read_vectored`
-    /// implementation, code using it may want to avoid the method all together
-    /// and coalesce writes into a single buffer for higher performance.
-    ///
-    /// The default implementation returns `false`.
-    #[unstable(feature = "can_vector", issue = "69941")]
-    fn is_read_vectored(&self) -> bool {
-        false
-    }
-
-    /// Reads all bytes until EOF in this source, placing them into `buf`.
-    ///
-    /// All bytes read from this source will be appended to the specified buffer
-    /// `buf`. This function will continuously call [`read()`] to append more data to
-    /// `buf` until [`read()`] returns either [`Ok(0)`] or an error of
-    /// non-[`ErrorKind::Interrupted`] kind.
-    ///
-    /// If successful, this function will return the total number of bytes read.
-    ///
-    /// # Errors
-    ///
-    /// If this function encounters an error of the kind
-    /// [`ErrorKind::Interrupted`] then the error is ignored and the operation
-    /// will continue.
-    ///
-    /// If any other read error is encountered then this function immediately
-    /// returns. Any bytes which have already been read will be appended to
-    /// `buf`.
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`read()`]: Read::read
-    /// [`Ok(0)`]: Ok
-    /// [`File`]: crate::fs::File
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///     let mut buffer = Vec::new();
-    ///
-    ///     // read the whole file
-    ///     f.read_to_end(&mut buffer)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    ///
-    /// (See also the [`std::fs::read`] convenience function for reading from a
-    /// file.)
-    ///
-    /// [`std::fs::read`]: crate::fs::read
-    ///
-    /// ## Implementing `read_to_end`
-    ///
-    /// When implementing the `io::Read` trait, it is recommended to allocate
-    /// memory using [`Vec::try_reserve`]. However, this behavior is not guaranteed
-    /// by all implementations, and `read_to_end` may not handle out-of-memory
-    /// situations gracefully.
-    ///
-    /// ```no_run
-    /// # use std::io::{self, BufRead};
-    /// # struct Example { example_datasource: io::Empty } impl Example {
-    /// # fn get_some_data_for_the_example(&self) -> &'static [u8] { &[] }
-    /// fn read_to_end(&mut self, dest_vec: &mut Vec<u8>) -> io::Result<usize> {
-    ///     let initial_vec_len = dest_vec.len();
-    ///     loop {
-    ///         let src_buf = self.example_datasource.fill_buf()?;
-    ///         if src_buf.is_empty() {
-    ///             break;
-    ///         }
-    ///         dest_vec.try_reserve(src_buf.len())?;
-    ///         dest_vec.extend_from_slice(src_buf);
-    ///
-    ///         // Any irreversible side effects should happen after `try_reserve` succeeds,
-    ///         // to avoid losing data on allocation error.
-    ///         let read = src_buf.len();
-    ///         self.example_datasource.consume(read);
-    ///     }
-    ///     Ok(dest_vec.len() - initial_vec_len)
-    /// }
-    /// # }
-    /// ```
-    ///
-    /// # Usage Notes
-    ///
-    /// `read_to_end` attempts to read a source until EOF, but many sources are continuous streams
-    /// that do not send EOF. In these cases, `read_to_end` will block indefinitely. Standard input
-    /// is one such stream which may be finite if piped, but is typically continuous. For example,
-    /// `cat file | my-rust-program` will correctly terminate with an `EOF` upon closure of cat.
-    /// Reading user input or running programs that remain open indefinitely will never terminate
-    /// the stream with `EOF` (e.g. `yes | my-rust-program`).
-    ///
-    /// Using `.lines()` with a [`BufReader`] or using [`read`] can provide a better solution
-    ///
-    ///[`read`]: Read::read
-    ///
-    /// [`Vec::try_reserve`]: crate::vec::Vec::try_reserve
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn read_to_end(&mut self, buf: &mut Vec<u8>) -> Result<usize> {
-        default_read_to_end(self, buf, None)
-    }
-
-    /// Reads all bytes until EOF in this source, appending them to `buf`.
-    ///
-    /// If successful, this function returns the number of bytes which were read
-    /// and appended to `buf`.
-    ///
-    /// # Errors
-    ///
-    /// If the data in this stream is *not* valid UTF-8 then an error is
-    /// returned and `buf` is unchanged.
-    ///
-    /// See [`read_to_end`] for other error semantics.
-    ///
-    /// [`read_to_end`]: Read::read_to_end
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`File`]: crate::fs::File
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///     let mut buffer = String::new();
-    ///
-    ///     f.read_to_string(&mut buffer)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    ///
-    /// (See also the [`std::fs::read_to_string`] convenience function for
-    /// reading from a file.)
-    ///
-    /// # Usage Notes
-    ///
-    /// `read_to_string` attempts to read a source until EOF, but many sources are continuous streams
-    /// that do not send EOF. In these cases, `read_to_string` will block indefinitely. Standard input
-    /// is one such stream which may be finite if piped, but is typically continuous. For example,
-    /// `cat file | my-rust-program` will correctly terminate with an `EOF` upon closure of cat.
-    /// Reading user input or running programs that remain open indefinitely will never terminate
-    /// the stream with `EOF` (e.g. `yes | my-rust-program`).
-    ///
-    /// Using `.lines()` with a [`BufReader`] or using [`read`] can provide a better solution
-    ///
-    ///[`read`]: Read::read
-    ///
-    /// [`std::fs::read_to_string`]: crate::fs::read_to_string
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn read_to_string(&mut self, buf: &mut String) -> Result<usize> {
-        default_read_to_string(self, buf, None)
-    }
-
-    /// Reads the exact number of bytes required to fill `buf`.
-    ///
-    /// This function reads as many bytes as necessary to completely fill the
-    /// specified buffer `buf`.
-    ///
-    /// *Implementations* of this method can make no assumptions about the contents of `buf` when
-    /// this function is called. It is recommended that implementations only write data to `buf`
-    /// instead of reading its contents. The documentation on [`read`] has a more detailed
-    /// explanation of this subject.
-    ///
-    /// # Errors
-    ///
-    /// If this function encounters an error of the kind
-    /// [`ErrorKind::Interrupted`] then the error is ignored and the operation
-    /// will continue.
-    ///
-    /// If this function encounters an "end of file" before completely filling
-    /// the buffer, it returns an error of the kind [`ErrorKind::UnexpectedEof`].
-    /// The contents of `buf` are unspecified in this case.
-    ///
-    /// If any other read error is encountered then this function immediately
-    /// returns. The contents of `buf` are unspecified in this case.
-    ///
-    /// If this function returns an error, it is unspecified how many bytes it
-    /// has read, but it will never read more than would be necessary to
-    /// completely fill the buffer.
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`read`]: Read::read
-    /// [`File`]: crate::fs::File
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///     let mut buffer = [0; 10];
-    ///
-    ///     // read exactly 10 bytes
-    ///     f.read_exact(&mut buffer)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "read_exact", since = "1.6.0")]
-    fn read_exact(&mut self, buf: &mut [u8]) -> Result<()> {
-        default_read_exact(self, buf)
-    }
-
-    /// Pull some bytes from this source into the specified buffer.
-    ///
-    /// This is equivalent to the [`read`](Read::read) method, except that it is passed a [`BorrowedCursor`] rather than `[u8]` to allow use
-    /// with uninitialized buffers. The new data will be appended to any existing contents of `buf`.
-    ///
-    /// The default implementation delegates to `read`.
-    ///
-    /// This method makes it possible to return both data and an error but it is advised against.
-    #[unstable(feature = "read_buf", issue = "78485")]
-    fn read_buf(&mut self, buf: BorrowedCursor<'_>) -> Result<()> {
-        default_read_buf(|b| self.read(b), buf)
-    }
-
-    /// Reads the exact number of bytes required to fill `cursor`.
-    ///
-    /// This is similar to the [`read_exact`](Read::read_exact) method, except
-    /// that it is passed a [`BorrowedCursor`] rather than `[u8]` to allow use
-    /// with uninitialized buffers.
-    ///
-    /// # Errors
-    ///
-    /// If this function encounters an error of the kind [`ErrorKind::Interrupted`]
-    /// then the error is ignored and the operation will continue.
-    ///
-    /// If this function encounters an "end of file" before completely filling
-    /// the buffer, it returns an error of the kind [`ErrorKind::UnexpectedEof`].
-    ///
-    /// If any other read error is encountered then this function immediately
-    /// returns.
-    ///
-    /// If this function returns an error, all bytes read will be appended to `cursor`.
-    #[unstable(feature = "read_buf", issue = "78485")]
-    fn read_buf_exact(&mut self, cursor: BorrowedCursor<'_>) -> Result<()> {
-        default_read_buf_exact(self, cursor)
-    }
-
-    /// Creates a "by reference" adapter for this instance of `Read`.
-    ///
-    /// The returned adapter also implements `Read` and will simply borrow this
-    /// current reader.
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`File`]: crate::fs::File
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::Read;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///     let mut buffer = Vec::new();
-    ///     let mut other_buffer = Vec::new();
-    ///
-    ///     {
-    ///         let reference = f.by_ref();
-    ///
-    ///         // read at most 5 bytes
-    ///         reference.take(5).read_to_end(&mut buffer)?;
-    ///
-    ///     } // drop our &mut reference so we can use f again
-    ///
-    ///     // original file still usable, read the rest
-    ///     f.read_to_end(&mut other_buffer)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn by_ref(&mut self) -> &mut Self
-    where
-        Self: Sized,
-    {
-        self
-    }
-
-    /// Transforms this `Read` instance to an [`Iterator`] over its bytes.
-    ///
-    /// The returned type implements [`Iterator`] where the [`Item`] is
-    /// <code>[Result]<[u8], [io::Error]></code>.
-    /// The yielded item is [`Ok`] if a byte was successfully read and [`Err`]
-    /// otherwise. EOF is mapped to returning [`None`] from this iterator.
-    ///
-    /// The default implementation calls `read` for each byte,
-    /// which can be very inefficient for data that's not in memory,
-    /// such as [`File`]. Consider using a [`BufReader`] in such cases.
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`Item`]: Iterator::Item
-    /// [`File`]: crate::fs::File "fs::File"
-    /// [Result]: crate::result::Result "Result"
-    /// [io::Error]: self::Error "io::Error"
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::io::BufReader;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let f = BufReader::new(File::open("foo.txt")?);
-    ///
-    ///     for byte in f.bytes() {
-    ///         println!("{}", byte?);
-    ///     }
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn bytes(self) -> Bytes<Self>
-    where
-        Self: Sized,
-    {
-        Bytes { inner: self }
-    }
-
-    /// Creates an adapter which will chain this stream with another.
-    ///
-    /// The returned `Read` instance will first read all bytes from this object
-    /// until EOF is encountered. Afterwards the output is equivalent to the
-    /// output of `next`.
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`File`]: crate::fs::File
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let f1 = File::open("foo.txt")?;
-    ///     let f2 = File::open("bar.txt")?;
-    ///
-    ///     let mut handle = f1.chain(f2);
-    ///     let mut buffer = String::new();
-    ///
-    ///     // read the value into a String. We could use any Read method here,
-    ///     // this is just one example.
-    ///     handle.read_to_string(&mut buffer)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn chain<R: Read>(self, next: R) -> Chain<Self, R>
-    where
-        Self: Sized,
-    {
-        Chain { first: self, second: next, done_first: false }
-    }
-
-    /// Creates an adapter which will read at most `limit` bytes from it.
-    ///
-    /// This function returns a new instance of `Read` which will read at most
-    /// `limit` bytes, after which it will always return EOF ([`Ok(0)`]). Any
-    /// read errors will not count towards the number of bytes read and future
-    /// calls to [`read()`] may succeed.
-    ///
-    /// # Examples
-    ///
-    /// [`File`]s implement `Read`:
-    ///
-    /// [`File`]: crate::fs::File
-    /// [`Ok(0)`]: Ok
-    /// [`read()`]: Read::read
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let f = File::open("foo.txt")?;
-    ///     let mut buffer = [0; 5];
-    ///
-    ///     // read at most five bytes
-    ///     let mut handle = f.take(5);
-    ///
-    ///     handle.read(&mut buffer)?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn take(self, limit: u64) -> Take<Self>
-    where
-        Self: Sized,
-    {
-        Take { inner: self, len: limit, limit }
-    }
-
-    /// Read and return a fixed array of bytes from this source.
-    ///
-    /// This function uses an array sized based on a const generic size known at compile time. You
-    /// can specify the size with turbofish (`reader.read_array::<8>()`), or let type inference
-    /// determine the number of bytes needed based on how the return value gets used. For instance,
-    /// this function works well with functions like [`u64::from_le_bytes`] to turn an array of
-    /// bytes into an integer of the same size.
-    ///
-    /// Like `read_exact`, if this function encounters an "end of file" before reading the desired
-    /// number of bytes, it returns an error of the kind [`ErrorKind::UnexpectedEof`].
-    ///
-    /// ```
-    /// #![feature(read_array)]
-    /// use std::io::Cursor;
-    /// use std::io::prelude::*;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let mut buf = Cursor::new([1, 2, 3, 4, 5, 6, 7, 8, 9, 8, 7, 6, 5, 4, 3, 2]);
-    ///     let x = u64::from_le_bytes(buf.read_array()?);
-    ///     let y = u32::from_be_bytes(buf.read_array()?);
-    ///     let z = u16::from_be_bytes(buf.read_array()?);
-    ///     assert_eq!(x, 0x807060504030201);
-    ///     assert_eq!(y, 0x9080706);
-    ///     assert_eq!(z, 0x504);
-    ///     Ok(())
-    /// }
-    /// ```
-    #[unstable(feature = "read_array", issue = "148848")]
-    fn read_array<const N: usize>(&mut self) -> Result<[u8; N]>
-    where
-        Self: Sized,
-    {
-        let mut buf = [MaybeUninit::uninit(); N];
-        let mut borrowed_buf = BorrowedBuf::from(buf.as_mut_slice());
-        self.read_buf_exact(borrowed_buf.unfilled())?;
-        // Guard against incorrect `read_buf_exact` implementations.
-        assert_eq!(borrowed_buf.len(), N);
-        Ok(unsafe { MaybeUninit::array_assume_init(buf) })
-    }
-}
-
-/// Reads all bytes from a [reader][Read] into a new [`String`].
-///
-/// This is a convenience function for [`Read::read_to_string`]. Using this
-/// function avoids having to create a variable first and provides more type
-/// safety since you can only get the buffer out if there were no errors. (If you
-/// use [`Read::read_to_string`] you have to remember to check whether the read
-/// succeeded because otherwise your buffer will be empty or only partially full.)
-///
-/// # Performance
-///
-/// The downside of this function's increased ease of use and type safety is
-/// that it gives you less control over performance. For example, you can't
-/// pre-allocate memory like you can using [`String::with_capacity`] and
-/// [`Read::read_to_string`]. Also, you can't re-use the buffer if an error
-/// occurs while reading.
-///
-/// In many cases, this function's performance will be adequate and the ease of use
-/// and type safety tradeoffs will be worth it. However, there are cases where you
-/// need more control over performance, and in those cases you should definitely use
-/// [`Read::read_to_string`] directly.
-///
-/// Note that in some special cases, such as when reading files, this function will
-/// pre-allocate memory based on the size of the input it is reading. In those
-/// cases, the performance should be as good as if you had used
-/// [`Read::read_to_string`] with a manually pre-allocated buffer.
-///
-/// # Errors
-///
-/// This function forces you to handle errors because the output (the `String`)
-/// is wrapped in a [`Result`]. See [`Read::read_to_string`] for the errors
-/// that can occur. If any error occurs, you will get an [`Err`], so you
-/// don't have to worry about your buffer being empty or partially full.
-///
-/// # Examples
-///
-/// ```no_run
-/// # use std::io;
-/// fn main() -> io::Result<()> {
-///     let stdin = io::read_to_string(io::stdin())?;
-///     println!("Stdin was:");
-///     println!("{stdin}");
-///     Ok(())
-/// }
-/// ```
-///
-/// # Usage Notes
-///
-/// `read_to_string` attempts to read a source until EOF, but many sources are continuous streams
-/// that do not send EOF. In these cases, `read_to_string` will block indefinitely. Standard input
-/// is one such stream which may be finite if piped, but is typically continuous. For example,
-/// `cat file | my-rust-program` will correctly terminate with an `EOF` upon closure of cat.
-/// Reading user input or running programs that remain open indefinitely will never terminate
-/// the stream with `EOF` (e.g. `yes | my-rust-program`).
-///
-/// Using `.lines()` with a [`BufReader`] or using [`read`] can provide a better solution
-///
-///[`read`]: Read::read
-///
-#[stable(feature = "io_read_to_string", since = "1.65.0")]
-pub fn read_to_string<R: Read>(mut reader: R) -> Result<String> {
-    let mut buf = String::new();
-    reader.read_to_string(&mut buf)?;
-    Ok(buf)
-}
-
-/// A trait for objects which are byte-oriented sinks.
-///
-/// Implementors of the `Write` trait are sometimes called 'writers'.
-///
-/// Writers are defined by two required methods, [`write`] and [`flush`]:
-///
-/// * The [`write`] method will attempt to write some data into the object,
-///   returning how many bytes were successfully written.
-///
-/// * The [`flush`] method is useful for adapters and explicit buffers
-///   themselves for ensuring that all buffered data has been pushed out to the
-///   'true sink'.
-///
-/// Writers are intended to be composable with one another. Many implementors
-/// throughout [`std::io`] take and provide types which implement the `Write`
-/// trait.
-///
-/// [`write`]: Write::write
-/// [`flush`]: Write::flush
-/// [`std::io`]: self
-///
-/// # Examples
-///
-/// ```no_run
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> std::io::Result<()> {
-///     let data = b"some bytes";
-///
-///     let mut pos = 0;
-///     let mut buffer = File::create("foo.txt")?;
-///
-///     while pos < data.len() {
-///         let bytes_written = buffer.write(&data[pos..])?;
-///         pos += bytes_written;
-///     }
-///     Ok(())
-/// }
-/// ```
-///
-/// The trait also provides convenience methods like [`write_all`], which calls
-/// `write` in a loop until its entire input has been written.
-///
-/// [`write_all`]: Write::write_all
-#[stable(feature = "rust1", since = "1.0.0")]
-#[doc(notable_trait)]
-#[cfg_attr(not(test), rustc_diagnostic_item = "IoWrite")]
-pub trait Write {
-    /// Writes a buffer into this writer, returning how many bytes were written.
-    ///
-    /// This function will attempt to write the entire contents of `buf`, but
-    /// the entire write might not succeed, or the write may also generate an
-    /// error. Typically, a call to `write` represents one attempt to write to
-    /// any wrapped object.
-    ///
-    /// Calls to `write` are not guaranteed to block waiting for data to be
-    /// written, and a write which would otherwise block can be indicated through
-    /// an [`Err`] variant.
-    ///
-    /// If this method consumed `n > 0` bytes of `buf` it must return [`Ok(n)`].
-    /// If the return value is `Ok(n)` then `n` must satisfy `n <= buf.len()`.
-    /// A return value of `Ok(0)` typically means that the underlying object is
-    /// no longer able to accept bytes and will likely not be able to in the
-    /// future as well, or that the buffer provided is empty.
-    ///
-    /// # Errors
-    ///
-    /// Each call to `write` may generate an I/O error indicating that the
-    /// operation could not be completed. If an error is returned then no bytes
-    /// in the buffer were written to this writer.
-    ///
-    /// It is **not** considered an error if the entire buffer could not be
-    /// written to this writer.
-    ///
-    /// An error of the [`ErrorKind::Interrupted`] kind is non-fatal and the
-    /// write operation should be retried if there is nothing else to do.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let mut buffer = File::create("foo.txt")?;
-    ///
-    ///     // Writes some prefix of the byte string, not necessarily all of it.
-    ///     buffer.write(b"some bytes")?;
-    ///     Ok(())
-    /// }
-    /// ```
-    ///
-    /// [`Ok(n)`]: Ok
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn write(&mut self, buf: &[u8]) -> Result<usize>;
-
-    /// Like [`write`], except that it writes from a slice of buffers.
-    ///
-    /// Data is copied from each buffer in order, with the final buffer
-    /// read from possibly being only partially consumed. This method must
-    /// behave as a call to [`write`] with the buffers concatenated would.
-    ///
-    /// The default implementation calls [`write`] with either the first nonempty
-    /// buffer provided, or an empty one if none exists.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io::IoSlice;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let data1 = [1; 8];
-    ///     let data2 = [15; 8];
-    ///     let io_slice1 = IoSlice::new(&data1);
-    ///     let io_slice2 = IoSlice::new(&data2);
-    ///
-    ///     let mut buffer = File::create("foo.txt")?;
-    ///
-    ///     // Writes some prefix of the byte string, not necessarily all of it.
-    ///     buffer.write_vectored(&[io_slice1, io_slice2])?;
-    ///     Ok(())
-    /// }
-    /// ```
-    ///
-    /// [`write`]: Write::write
-    #[stable(feature = "iovec", since = "1.36.0")]
-    fn write_vectored(&mut self, bufs: &[IoSlice<'_>]) -> Result<usize> {
-        default_write_vectored(|b| self.write(b), bufs)
-    }
-
-    /// Determines if this `Write`r has an efficient [`write_vectored`]
-    /// implementation.
-    ///
-    /// If a `Write`r does not override the default [`write_vectored`]
-    /// implementation, code using it may want to avoid the method all together
-    /// and coalesce writes into a single buffer for higher performance.
-    ///
-    /// The default implementation returns `false`.
-    ///
-    /// [`write_vectored`]: Write::write_vectored
-    #[unstable(feature = "can_vector", issue = "69941")]
-    fn is_write_vectored(&self) -> bool {
-        false
-    }
-
-    /// Flushes this output stream, ensuring that all intermediately buffered
-    /// contents reach their destination.
-    ///
-    /// # Errors
-    ///
-    /// It is considered an error if not all bytes could be written due to
-    /// I/O errors or EOF being reached.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io::prelude::*;
-    /// use std::io::BufWriter;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let mut buffer = BufWriter::new(File::create("foo.txt")?);
-    ///
-    ///     buffer.write_all(b"some bytes")?;
-    ///     buffer.flush()?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn flush(&mut self) -> Result<()>;
-
-    /// Attempts to write an entire buffer into this writer.
-    ///
-    /// This method will continuously call [`write`] until there is no more data
-    /// to be written or an error of non-[`ErrorKind::Interrupted`] kind is
-    /// returned. This method will not return until the entire buffer has been
-    /// successfully written or such an error occurs. The first error that is
-    /// not of [`ErrorKind::Interrupted`] kind generated from this method will be
-    /// returned.
-    ///
-    /// If the buffer contains no data, this will never call [`write`].
-    ///
-    /// # Errors
-    ///
-    /// This function will return the first error of
-    /// non-[`ErrorKind::Interrupted`] kind that [`write`] returns.
-    ///
-    /// [`write`]: Write::write
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let mut buffer = File::create("foo.txt")?;
-    ///
-    ///     buffer.write_all(b"some bytes")?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn write_all(&mut self, mut buf: &[u8]) -> Result<()> {
-        while !buf.is_empty() {
-            match self.write(buf) {
-                Ok(0) => {
-                    return Err(Error::WRITE_ALL_EOF);
-                }
-                Ok(n) => buf = &buf[n..],
-                Err(ref e) if e.is_interrupted() => {}
-                Err(e) => return Err(e),
-            }
-        }
-        Ok(())
-    }
-
-    /// Attempts to write multiple buffers into this writer.
-    ///
-    /// This method will continuously call [`write_vectored`] until there is no
-    /// more data to be written or an error of non-[`ErrorKind::Interrupted`]
-    /// kind is returned. This method will not return until all buffers have
-    /// been successfully written or such an error occurs. The first error that
-    /// is not of [`ErrorKind::Interrupted`] kind generated from this method
-    /// will be returned.
-    ///
-    /// If the buffer contains no data, this will never call [`write_vectored`].
-    ///
-    /// # Notes
-    ///
-    /// Unlike [`write_vectored`], this takes a *mutable* reference to
-    /// a slice of [`IoSlice`]s, not an immutable one. That's because we need to
-    /// modify the slice to keep track of the bytes already written.
-    ///
-    /// Once this function returns, the contents of `bufs` are unspecified, as
-    /// this depends on how many calls to [`write_vectored`] were necessary. It is
-    /// best to understand this function as taking ownership of `bufs` and to
-    /// not use `bufs` afterwards. The underlying buffers, to which the
-    /// [`IoSlice`]s point (but not the [`IoSlice`]s themselves), are unchanged and
-    /// can be reused.
-    ///
-    /// [`write_vectored`]: Write::write_vectored
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// #![feature(write_all_vectored)]
-    /// # fn main() -> std::io::Result<()> {
-    ///
-    /// use std::io::{Write, IoSlice};
-    ///
-    /// let mut writer = Vec::new();
-    /// let bufs = &mut [
-    ///     IoSlice::new(&[1]),
-    ///     IoSlice::new(&[2, 3]),
-    ///     IoSlice::new(&[4, 5, 6]),
-    /// ];
-    ///
-    /// writer.write_all_vectored(bufs)?;
-    /// // Note: the contents of `bufs` is now undefined, see the Notes section.
-    ///
-    /// assert_eq!(writer, &[1, 2, 3, 4, 5, 6]);
-    /// # Ok(()) }
-    /// ```
-    #[unstable(feature = "write_all_vectored", issue = "70436")]
-    fn write_all_vectored(&mut self, mut bufs: &mut [IoSlice<'_>]) -> Result<()> {
-        // Guarantee that bufs is empty if it contains no data,
-        // to avoid calling write_vectored if there is no data to be written.
-        IoSlice::advance_slices(&mut bufs, 0);
-        while !bufs.is_empty() {
-            match self.write_vectored(bufs) {
-                Ok(0) => {
-                    return Err(Error::WRITE_ALL_EOF);
-                }
-                Ok(n) => IoSlice::advance_slices(&mut bufs, n),
-                Err(ref e) if e.is_interrupted() => {}
-                Err(e) => return Err(e),
-            }
-        }
-        Ok(())
-    }
-
-    /// Writes a formatted string into this writer, returning any error
-    /// encountered.
-    ///
-    /// This method is primarily used to interface with the
-    /// [`format_args!()`] macro, and it is rare that this should
-    /// explicitly be called. The [`write!()`] macro should be favored to
-    /// invoke this method instead.
-    ///
-    /// This function internally uses the [`write_all`] method on
-    /// this trait and hence will continuously write data so long as no errors
-    /// are received. This also means that partial writes are not indicated in
-    /// this signature.
-    ///
-    /// [`write_all`]: Write::write_all
-    ///
-    /// # Errors
-    ///
-    /// This function will return any I/O error reported while formatting.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let mut buffer = File::create("foo.txt")?;
-    ///
-    ///     // this call
-    ///     write!(buffer, "{:.*}", 2, 1.234567)?;
-    ///     // turns into this:
-    ///     buffer.write_fmt(format_args!("{:.*}", 2, 1.234567))?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn write_fmt(&mut self, args: fmt::Arguments<'_>) -> Result<()> {
-        if let Some(s) = args.as_statically_known_str() {
-            self.write_all(s.as_bytes())
-        } else {
-            default_write_fmt(self, args)
-        }
-    }
-
-    /// Creates a "by reference" adapter for this instance of `Write`.
-    ///
-    /// The returned adapter also implements `Write` and will simply borrow this
-    /// current writer.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io::Write;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> std::io::Result<()> {
-    ///     let mut buffer = File::create("foo.txt")?;
-    ///
-    ///     let reference = buffer.by_ref();
-    ///
-    ///     // we can use reference just like our original buffer
-    ///     reference.write_all(b"some bytes")?;
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn by_ref(&mut self) -> &mut Self
-    where
-        Self: Sized,
-    {
-        self
-    }
-}
-
-/// The `Seek` trait provides a cursor which can be moved within a stream of
-/// bytes.
-///
-/// The stream typically has a fixed size, allowing seeking relative to either
-/// end or the current offset.
-///
-/// # Examples
-///
-/// [`File`]s implement `Seek`:
-///
-/// [`File`]: crate::fs::File
-///
-/// ```no_run
-/// use std::io;
-/// use std::io::prelude::*;
-/// use std::fs::File;
-/// use std::io::SeekFrom;
-///
-/// fn main() -> io::Result<()> {
-///     let mut f = File::open("foo.txt")?;
-///
-///     // move the cursor 42 bytes from the start of the file
-///     f.seek(SeekFrom::Start(42))?;
-///     Ok(())
-/// }
-/// ```
-#[stable(feature = "rust1", since = "1.0.0")]
-#[cfg_attr(not(test), rustc_diagnostic_item = "IoSeek")]
-pub trait Seek {
-    /// Seek to an offset, in bytes, in a stream.
-    ///
-    /// A seek beyond the end of a stream is allowed, but behavior is defined
-    /// by the implementation.
-    ///
-    /// If the seek operation completed successfully,
-    /// this method returns the new position from the start of the stream.
-    /// That position can be used later with [`SeekFrom::Start`].
-    ///
-    /// # Errors
-    ///
-    /// Seeking can fail, for example because it might involve flushing a buffer.
-    ///
-    /// Seeking to a negative offset is considered an error.
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn seek(&mut self, pos: SeekFrom) -> Result<u64>;
-
-    /// Rewind to the beginning of a stream.
-    ///
-    /// This is a convenience method, equivalent to `seek(SeekFrom::Start(0))`.
-    ///
-    /// # Errors
-    ///
-    /// Rewinding can fail, for example because it might involve flushing a buffer.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use std::io::{Read, Seek, Write};
-    /// use std::fs::OpenOptions;
-    ///
-    /// let mut f = OpenOptions::new()
-    ///     .write(true)
-    ///     .read(true)
-    ///     .create(true)
-    ///     .open("foo.txt")?;
-    ///
-    /// let hello = "Hello!\n";
-    /// write!(f, "{hello}")?;
-    /// f.rewind()?;
-    ///
-    /// let mut buf = String::new();
-    /// f.read_to_string(&mut buf)?;
-    /// assert_eq!(&buf, hello);
-    /// # std::io::Result::Ok(())
-    /// ```
-    #[stable(feature = "seek_rewind", since = "1.55.0")]
-    fn rewind(&mut self) -> Result<()> {
-        self.seek(SeekFrom::Start(0))?;
-        Ok(())
-    }
-
-    /// Returns the length of this stream (in bytes).
-    ///
-    /// The default implementation uses up to three seek operations. If this
-    /// method returns successfully, the seek position is unchanged (i.e. the
-    /// position before calling this method is the same as afterwards).
-    /// However, if this method returns an error, the seek position is
-    /// unspecified.
-    ///
-    /// If you need to obtain the length of *many* streams and you don't care
-    /// about the seek position afterwards, you can reduce the number of seek
-    /// operations by simply calling `seek(SeekFrom::End(0))` and using its
-    /// return value (it is also the stream length).
-    ///
-    /// Note that length of a stream can change over time (for example, when
-    /// data is appended to a file). So calling this method multiple times does
-    /// not necessarily return the same length each time.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// #![feature(seek_stream_len)]
-    /// use std::{
-    ///     io::{self, Seek},
-    ///     fs::File,
-    /// };
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///
-    ///     let len = f.stream_len()?;
-    ///     println!("The file is currently {len} bytes long");
-    ///     Ok(())
-    /// }
-    /// ```
-    #[unstable(feature = "seek_stream_len", issue = "59359")]
-    fn stream_len(&mut self) -> Result<u64> {
-        stream_len_default(self)
-    }
-
-    /// Returns the current seek position from the start of the stream.
-    ///
-    /// This is equivalent to `self.seek(SeekFrom::Current(0))`.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use std::{
-    ///     io::{self, BufRead, BufReader, Seek},
-    ///     fs::File,
-    /// };
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = BufReader::new(File::open("foo.txt")?);
-    ///
-    ///     let before = f.stream_position()?;
-    ///     f.read_line(&mut String::new())?;
-    ///     let after = f.stream_position()?;
-    ///
-    ///     println!("The first line was {} bytes long", after - before);
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "seek_convenience", since = "1.51.0")]
-    fn stream_position(&mut self) -> Result<u64> {
-        self.seek(SeekFrom::Current(0))
-    }
-
-    /// Seeks relative to the current position.
-    ///
-    /// This is equivalent to `self.seek(SeekFrom::Current(offset))` but
-    /// doesn't return the new position which can allow some implementations
-    /// such as [`BufReader`] to perform more efficient seeks.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// use std::{
-    ///     io::{self, Seek},
-    ///     fs::File,
-    /// };
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut f = File::open("foo.txt")?;
-    ///     f.seek_relative(10)?;
-    ///     assert_eq!(f.stream_position()?, 10);
-    ///     Ok(())
-    /// }
-    /// ```
-    ///
-    /// [`BufReader`]: crate::io::BufReader
-    #[stable(feature = "seek_seek_relative", since = "1.80.0")]
-    fn seek_relative(&mut self, offset: i64) -> Result<()> {
-        self.seek(SeekFrom::Current(offset))?;
-        Ok(())
-    }
-}
-
-pub(crate) fn stream_len_default<T: Seek + ?Sized>(self_: &mut T) -> Result<u64> {
-    let old_pos = self_.stream_position()?;
-    let len = self_.seek(SeekFrom::End(0))?;
-
-    // Avoid seeking a third time when we were already at the end of the
-    // stream. The branch is usually way cheaper than a seek operation.
-    if old_pos != len {
-        self_.seek(SeekFrom::Start(old_pos))?;
-    }
-
-    Ok(len)
-}
-
-/// Enumeration of possible methods to seek within an I/O object.
-///
-/// It is used by the [`Seek`] trait.
-#[derive(Copy, PartialEq, Eq, Clone, Debug)]
-#[stable(feature = "rust1", since = "1.0.0")]
-#[cfg_attr(not(test), rustc_diagnostic_item = "SeekFrom")]
-pub enum SeekFrom {
-    /// Sets the offset to the provided number of bytes.
-    #[stable(feature = "rust1", since = "1.0.0")]
-    Start(#[stable(feature = "rust1", since = "1.0.0")] u64),
-
-    /// Sets the offset to the size of this object plus the specified number of
-    /// bytes.
-    ///
-    /// It is possible to seek beyond the end of an object, but it's an error to
-    /// seek before byte 0.
-    #[stable(feature = "rust1", since = "1.0.0")]
-    End(#[stable(feature = "rust1", since = "1.0.0")] i64),
-
-    /// Sets the offset to the current position plus the specified number of
-    /// bytes.
-    ///
-    /// It is possible to seek beyond the end of an object, but it's an error to
-    /// seek before byte 0.
-    #[stable(feature = "rust1", since = "1.0.0")]
-    Current(#[stable(feature = "rust1", since = "1.0.0")] i64),
-}
-
-fn read_until<R: BufRead + ?Sized>(r: &mut R, delim: u8, buf: &mut Vec<u8>) -> Result<usize> {
-    let mut read = 0;
-    loop {
-        let (done, used) = {
-            let available = match r.fill_buf() {
-                Ok(n) => n,
-                Err(ref e) if e.is_interrupted() => continue,
-                Err(e) => return Err(e),
-            };
-            match memchr::memchr(delim, available) {
-                Some(i) => {
-                    buf.extend_from_slice(&available[..=i]);
-                    (true, i + 1)
-                }
-                None => {
-                    buf.extend_from_slice(available);
-                    (false, available.len())
-                }
-            }
-        };
-        r.consume(used);
-        read += used;
-        if done || used == 0 {
-            return Ok(read);
-        }
-    }
-}
-
-fn skip_until<R: BufRead + ?Sized>(r: &mut R, delim: u8) -> Result<usize> {
-    let mut read = 0;
-    loop {
-        let (done, used) = {
-            let available = match r.fill_buf() {
-                Ok(n) => n,
-                Err(ref e) if e.kind() == ErrorKind::Interrupted => continue,
-                Err(e) => return Err(e),
-            };
-            match memchr::memchr(delim, available) {
-                Some(i) => (true, i + 1),
-                None => (false, available.len()),
-            }
-        };
-        r.consume(used);
-        read += used;
-        if done || used == 0 {
-            return Ok(read);
-        }
-    }
-}
-
-/// A `BufRead` is a type of `Read`er which has an internal buffer, allowing it
-/// to perform extra ways of reading.
-///
-/// For example, reading line-by-line is inefficient without using a buffer, so
-/// if you want to read by line, you'll need `BufRead`, which includes a
-/// [`read_line`] method as well as a [`lines`] iterator.
-///
-/// # Examples
-///
-/// A locked standard input implements `BufRead`:
-///
-/// ```no_run
-/// use std::io;
-/// use std::io::prelude::*;
-///
-/// let stdin = io::stdin();
-/// for line in stdin.lock().lines() {
-///     println!("{}", line?);
-/// }
-/// # std::io::Result::Ok(())
-/// ```
-///
-/// If you have something that implements [`Read`], you can use the [`BufReader`
-/// type][`BufReader`] to turn it into a `BufRead`.
-///
-/// For example, [`File`] implements [`Read`], but not `BufRead`.
-/// [`BufReader`] to the rescue!
-///
-/// [`File`]: crate::fs::File
-/// [`read_line`]: BufRead::read_line
-/// [`lines`]: BufRead::lines
-///
-/// ```no_run
-/// use std::io::{self, BufReader};
-/// use std::io::prelude::*;
-/// use std::fs::File;
-///
-/// fn main() -> io::Result<()> {
-///     let f = File::open("foo.txt")?;
-///     let f = BufReader::new(f);
-///
-///     for line in f.lines() {
-///         let line = line?;
-///         println!("{line}");
-///     }
-///
-///     Ok(())
-/// }
-/// ```
-#[stable(feature = "rust1", since = "1.0.0")]
-#[cfg_attr(not(test), rustc_diagnostic_item = "IoBufRead")]
-pub trait BufRead: Read {
-    /// Returns the contents of the internal buffer, filling it with more data, via `Read` methods, if empty.
-    ///
-    /// This is a lower-level method and is meant to be used together with [`consume`],
-    /// which can be used to mark bytes that should not be returned by subsequent calls to `read`.
-    ///
-    /// [`consume`]: BufRead::consume
-    ///
-    /// Returns an empty buffer when the stream has reached EOF.
-    ///
-    /// # Errors
-    ///
-    /// This function will return an I/O error if a `Read` method was called, but returned an error.
-    ///
-    /// # Examples
-    ///
-    /// A locked standard input implements `BufRead`:
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    ///
-    /// let stdin = io::stdin();
-    /// let mut stdin = stdin.lock();
-    ///
-    /// let buffer = stdin.fill_buf()?;
-    ///
-    /// // work with buffer
-    /// println!("{buffer:?}");
-    ///
-    /// // mark the bytes we worked with as read
-    /// let length = buffer.len();
-    /// stdin.consume(length);
-    /// # std::io::Result::Ok(())
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn fill_buf(&mut self) -> Result<&[u8]>;
-
-    /// Marks the given `amount` of additional bytes from the internal buffer as having been read.
-    /// Subsequent calls to `read` only return bytes that have not been marked as read.
-    ///
-    /// This is a lower-level method and is meant to be used together with [`fill_buf`],
-    /// which can be used to fill the internal buffer via `Read` methods.
-    ///
-    /// It is a logic error if `amount` exceeds the number of unread bytes in the internal buffer, which is returned by [`fill_buf`].
-    ///
-    /// # Examples
-    ///
-    /// Since `consume()` is meant to be used with [`fill_buf`],
-    /// that method's example includes an example of `consume()`.
-    ///
-    /// [`fill_buf`]: BufRead::fill_buf
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn consume(&mut self, amount: usize);
-
-    /// Checks if there is any data left to be `read`.
-    ///
-    /// This function may fill the buffer to check for data,
-    /// so this function returns `Result<bool>`, not `bool`.
-    ///
-    /// The default implementation calls `fill_buf` and checks that the
-    /// returned slice is empty (which means that there is no data left,
-    /// since EOF is reached).
-    ///
-    /// # Errors
-    ///
-    /// This function will return an I/O error if a `Read` method was called, but returned an error.
-    ///
-    /// Examples
-    ///
-    /// ```
-    /// #![feature(buf_read_has_data_left)]
-    /// use std::io;
-    /// use std::io::prelude::*;
-    ///
-    /// let stdin = io::stdin();
-    /// let mut stdin = stdin.lock();
-    ///
-    /// while stdin.has_data_left()? {
-    ///     let mut line = String::new();
-    ///     stdin.read_line(&mut line)?;
-    ///     // work with line
-    ///     println!("{line:?}");
-    /// }
-    /// # std::io::Result::Ok(())
-    /// ```
-    #[unstable(feature = "buf_read_has_data_left", issue = "86423")]
-    fn has_data_left(&mut self) -> Result<bool> {
-        self.fill_buf().map(|b| !b.is_empty())
-    }
-
-    /// Reads all bytes into `buf` until the delimiter `byte` or EOF is reached.
-    ///
-    /// This function will read bytes from the underlying stream until the
-    /// delimiter or EOF is found. Once found, all bytes up to, and including,
-    /// the delimiter (if found) will be appended to `buf`.
-    ///
-    /// If successful, this function will return the total number of bytes read.
-    ///
-    /// This function is blocking and should be used carefully: it is possible for
-    /// an attacker to continuously send bytes without ever sending the delimiter
-    /// or EOF.
-    ///
-    /// # Errors
-    ///
-    /// This function will ignore all instances of [`ErrorKind::Interrupted`] and
-    /// will otherwise return any errors returned by [`fill_buf`].
-    ///
-    /// If an I/O error is encountered then all bytes read so far will be
-    /// present in `buf` and its length will have been adjusted appropriately.
-    ///
-    /// [`fill_buf`]: BufRead::fill_buf
-    ///
-    /// # Examples
-    ///
-    /// [`std::io::Cursor`][`Cursor`] is a type that implements `BufRead`. In
-    /// this example, we use [`Cursor`] to read all the bytes in a byte slice
-    /// in hyphen delimited segments:
-    ///
-    /// ```
-    /// use std::io::{self, BufRead};
-    ///
-    /// let mut cursor = io::Cursor::new(b"lorem-ipsum");
-    /// let mut buf = vec![];
-    ///
-    /// // cursor is at 'l'
-    /// let num_bytes = cursor.read_until(b'-', &mut buf)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 6);
-    /// assert_eq!(buf, b"lorem-");
-    /// buf.clear();
-    ///
-    /// // cursor is at 'i'
-    /// let num_bytes = cursor.read_until(b'-', &mut buf)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 5);
-    /// assert_eq!(buf, b"ipsum");
-    /// buf.clear();
-    ///
-    /// // cursor is at EOF
-    /// let num_bytes = cursor.read_until(b'-', &mut buf)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 0);
-    /// assert_eq!(buf, b"");
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn read_until(&mut self, byte: u8, buf: &mut Vec<u8>) -> Result<usize> {
-        read_until(self, byte, buf)
-    }
-
-    /// Skips all bytes until the delimiter `byte` or EOF is reached.
-    ///
-    /// This function will read (and discard) bytes from the underlying stream until the
-    /// delimiter or EOF is found.
-    ///
-    /// If successful, this function will return the total number of bytes read,
-    /// including the delimiter byte if found.
-    ///
-    /// This is useful for efficiently skipping data such as NUL-terminated strings
-    /// in binary file formats without buffering.
-    ///
-    /// This function is blocking and should be used carefully: it is possible for
-    /// an attacker to continuously send bytes without ever sending the delimiter
-    /// or EOF.
-    ///
-    /// # Errors
-    ///
-    /// This function will ignore all instances of [`ErrorKind::Interrupted`] and
-    /// will otherwise return any errors returned by [`fill_buf`].
-    ///
-    /// If an I/O error is encountered then all bytes read so far will be
-    /// present in `buf` and its length will have been adjusted appropriately.
-    ///
-    /// [`fill_buf`]: BufRead::fill_buf
-    ///
-    /// # Examples
-    ///
-    /// [`std::io::Cursor`][`Cursor`] is a type that implements `BufRead`. In
-    /// this example, we use [`Cursor`] to read some NUL-terminated information
-    /// about Ferris from a binary string, skipping the fun fact:
-    ///
-    /// ```
-    /// use std::io::{self, BufRead};
-    ///
-    /// let mut cursor = io::Cursor::new(b"Ferris\0Likes long walks on the beach\0Crustacean\0!");
-    ///
-    /// // read name
-    /// let mut name = Vec::new();
-    /// let num_bytes = cursor.read_until(b'\0', &mut name)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 7);
-    /// assert_eq!(name, b"Ferris\0");
-    ///
-    /// // skip fun fact
-    /// let num_bytes = cursor.skip_until(b'\0')
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 30);
-    ///
-    /// // read animal type
-    /// let mut animal = Vec::new();
-    /// let num_bytes = cursor.read_until(b'\0', &mut animal)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 11);
-    /// assert_eq!(animal, b"Crustacean\0");
-    ///
-    /// // reach EOF
-    /// let num_bytes = cursor.skip_until(b'\0')
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 1);
-    /// ```
-    #[stable(feature = "bufread_skip_until", since = "1.83.0")]
-    fn skip_until(&mut self, byte: u8) -> Result<usize> {
-        skip_until(self, byte)
-    }
-
-    /// Reads all bytes until a newline (the `0xA` byte) is reached, and append
-    /// them to the provided `String` buffer.
-    ///
-    /// Previous content of the buffer will be preserved. To avoid appending to
-    /// the buffer, you need to [`clear`] it first.
-    ///
-    /// This function will read bytes from the underlying stream until the
-    /// newline delimiter (the `0xA` byte) or EOF is found. Once found, all bytes
-    /// up to, and including, the delimiter (if found) will be appended to
-    /// `buf`.
-    ///
-    /// If successful, this function will return the total number of bytes read.
-    ///
-    /// If this function returns [`Ok(0)`], the stream has reached EOF.
-    ///
-    /// This function is blocking and should be used carefully: it is possible for
-    /// an attacker to continuously send bytes without ever sending a newline
-    /// or EOF. You can use [`take`] to limit the maximum number of bytes read.
-    ///
-    /// [`Ok(0)`]: Ok
-    /// [`clear`]: String::clear
-    /// [`take`]: crate::io::Read::take
-    ///
-    /// # Errors
-    ///
-    /// This function has the same error semantics as [`read_until`] and will
-    /// also return an error if the read bytes are not valid UTF-8. If an I/O
-    /// error is encountered then `buf` may contain some bytes already read in
-    /// the event that all data read so far was valid UTF-8.
-    ///
-    /// [`read_until`]: BufRead::read_until
-    ///
-    /// # Examples
-    ///
-    /// [`std::io::Cursor`][`Cursor`] is a type that implements `BufRead`. In
-    /// this example, we use [`Cursor`] to read all the lines in a byte slice:
-    ///
-    /// ```
-    /// use std::io::{self, BufRead};
-    ///
-    /// let mut cursor = io::Cursor::new(b"foo\nbar");
-    /// let mut buf = String::new();
-    ///
-    /// // cursor is at 'f'
-    /// let num_bytes = cursor.read_line(&mut buf)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 4);
-    /// assert_eq!(buf, "foo\n");
-    /// buf.clear();
-    ///
-    /// // cursor is at 'b'
-    /// let num_bytes = cursor.read_line(&mut buf)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 3);
-    /// assert_eq!(buf, "bar");
-    /// buf.clear();
-    ///
-    /// // cursor is at EOF
-    /// let num_bytes = cursor.read_line(&mut buf)
-    ///     .expect("reading from cursor won't fail");
-    /// assert_eq!(num_bytes, 0);
-    /// assert_eq!(buf, "");
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn read_line(&mut self, buf: &mut String) -> Result<usize> {
-        // Note that we are not calling the `.read_until` method here, but
-        // rather our hardcoded implementation. For more details as to why, see
-        // the comments in `default_read_to_string`.
-        unsafe { append_to_string(buf, |b| read_until(self, b'\n', b)) }
-    }
-
-    /// Returns an iterator over the contents of this reader split on the byte
-    /// `byte`.
-    ///
-    /// The iterator returned from this function will return instances of
-    /// <code>[io::Result]<[Vec]\<u8>></code>. Each vector returned will *not* have
-    /// the delimiter byte at the end.
-    ///
-    /// This function will yield errors whenever [`read_until`] would have
-    /// also yielded an error.
-    ///
-    /// [io::Result]: self::Result "io::Result"
-    /// [`read_until`]: BufRead::read_until
-    ///
-    /// # Examples
-    ///
-    /// [`std::io::Cursor`][`Cursor`] is a type that implements `BufRead`. In
-    /// this example, we use [`Cursor`] to iterate over all hyphen delimited
-    /// segments in a byte slice
-    ///
-    /// ```
-    /// use std::io::{self, BufRead};
-    ///
-    /// let cursor = io::Cursor::new(b"lorem-ipsum-dolor");
-    ///
-    /// let mut split_iter = cursor.split(b'-').map(|l| l.unwrap());
-    /// assert_eq!(split_iter.next(), Some(b"lorem".to_vec()));
-    /// assert_eq!(split_iter.next(), Some(b"ipsum".to_vec()));
-    /// assert_eq!(split_iter.next(), Some(b"dolor".to_vec()));
-    /// assert_eq!(split_iter.next(), None);
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn split(self, byte: u8) -> Split<Self>
-    where
-        Self: Sized,
-    {
-        Split { buf: self, delim: byte }
-    }
-
-    /// Returns an iterator over the lines of this reader.
-    ///
-    /// The iterator returned from this function will yield instances of
-    /// <code>[io::Result]<[String]></code>. Each string returned will *not* have a newline
-    /// byte (the `0xA` byte) or `CRLF` (`0xD`, `0xA` bytes) at the end.
-    ///
-    /// [io::Result]: self::Result "io::Result"
-    ///
-    /// # Examples
-    ///
-    /// [`std::io::Cursor`][`Cursor`] is a type that implements `BufRead`. In
-    /// this example, we use [`Cursor`] to iterate over all the lines in a byte
-    /// slice.
-    ///
-    /// ```
-    /// use std::io::{self, BufRead};
-    ///
-    /// let cursor = io::Cursor::new(b"lorem\nipsum\r\ndolor");
-    ///
-    /// let mut lines_iter = cursor.lines().map(|l| l.unwrap());
-    /// assert_eq!(lines_iter.next(), Some(String::from("lorem")));
-    /// assert_eq!(lines_iter.next(), Some(String::from("ipsum")));
-    /// assert_eq!(lines_iter.next(), Some(String::from("dolor")));
-    /// assert_eq!(lines_iter.next(), None);
-    /// ```
-    ///
-    /// # Errors
-    ///
-    /// Each line of the iterator has the same error semantics as [`BufRead::read_line`].
-    #[stable(feature = "rust1", since = "1.0.0")]
-    fn lines(self) -> Lines<Self>
-    where
-        Self: Sized,
-    {
-        Lines { buf: self }
-    }
-}
-
-/// Adapter to chain together two readers.
-///
-/// This struct is generally created by calling [`chain`] on a reader.
-/// Please see the documentation of [`chain`] for more details.
-///
-/// [`chain`]: Read::chain
-#[stable(feature = "rust1", since = "1.0.0")]
-#[derive(Debug)]
-pub struct Chain<T, U> {
-    first: T,
-    second: U,
-    done_first: bool,
-}
-
-impl<T, U> Chain<T, U> {
-    /// Consumes the `Chain`, returning the wrapped readers.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut foo_file = File::open("foo.txt")?;
-    ///     let mut bar_file = File::open("bar.txt")?;
-    ///
-    ///     let chain = foo_file.chain(bar_file);
-    ///     let (foo_file, bar_file) = chain.into_inner();
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "more_io_inner_methods", since = "1.20.0")]
-    pub fn into_inner(self) -> (T, U) {
-        (self.first, self.second)
-    }
-
-    /// Gets references to the underlying readers in this `Chain`.
-    ///
-    /// Care should be taken to avoid modifying the internal I/O state of the
-    /// underlying readers as doing so may corrupt the internal state of this
-    /// `Chain`.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut foo_file = File::open("foo.txt")?;
-    ///     let mut bar_file = File::open("bar.txt")?;
-    ///
-    ///     let chain = foo_file.chain(bar_file);
-    ///     let (foo_file, bar_file) = chain.get_ref();
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "more_io_inner_methods", since = "1.20.0")]
-    pub fn get_ref(&self) -> (&T, &U) {
-        (&self.first, &self.second)
-    }
-
-    /// Gets mutable references to the underlying readers in this `Chain`.
-    ///
-    /// Care should be taken to avoid modifying the internal I/O state of the
-    /// underlying readers as doing so may corrupt the internal state of this
-    /// `Chain`.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut foo_file = File::open("foo.txt")?;
-    ///     let mut bar_file = File::open("bar.txt")?;
-    ///
-    ///     let mut chain = foo_file.chain(bar_file);
-    ///     let (foo_file, bar_file) = chain.get_mut();
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "more_io_inner_methods", since = "1.20.0")]
-    pub fn get_mut(&mut self) -> (&mut T, &mut U) {
-        (&mut self.first, &mut self.second)
-    }
-}
-
-#[stable(feature = "rust1", since = "1.0.0")]
-impl<T: Read, U: Read> Read for Chain<T, U> {
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize> {
-        if !self.done_first {
-            match self.first.read(buf)? {
-                0 if !buf.is_empty() => self.done_first = true,
-                n => return Ok(n),
-            }
-        }
-        self.second.read(buf)
-    }
-
-    fn read_vectored(&mut self, bufs: &mut [IoSliceMut<'_>]) -> Result<usize> {
-        if !self.done_first {
-            match self.first.read_vectored(bufs)? {
-                0 if bufs.iter().any(|b| !b.is_empty()) => self.done_first = true,
-                n => return Ok(n),
-            }
-        }
-        self.second.read_vectored(bufs)
-    }
-
-    #[inline]
-    fn is_read_vectored(&self) -> bool {
-        self.first.is_read_vectored() || self.second.is_read_vectored()
-    }
-
-    fn read_to_end(&mut self, buf: &mut Vec<u8>) -> Result<usize> {
-        let mut read = 0;
-        if !self.done_first {
-            read += self.first.read_to_end(buf)?;
-            self.done_first = true;
-        }
-        read += self.second.read_to_end(buf)?;
-        Ok(read)
-    }
-
-    // We don't override `read_to_string` here because an UTF-8 sequence could
-    // be split between the two parts of the chain
-
-    fn read_buf(&mut self, mut buf: BorrowedCursor<'_>) -> Result<()> {
-        if buf.capacity() == 0 {
-            return Ok(());
-        }
-
-        if !self.done_first {
-            let old_len = buf.written();
-            self.first.read_buf(buf.reborrow())?;
-
-            if buf.written() != old_len {
-                return Ok(());
-            } else {
-                self.done_first = true;
-            }
-        }
-        self.second.read_buf(buf)
-    }
-}
-
-#[stable(feature = "chain_bufread", since = "1.9.0")]
-impl<T: BufRead, U: BufRead> BufRead for Chain<T, U> {
-    fn fill_buf(&mut self) -> Result<&[u8]> {
-        if !self.done_first {
-            match self.first.fill_buf()? {
-                buf if buf.is_empty() => self.done_first = true,
-                buf => return Ok(buf),
-            }
-        }
-        self.second.fill_buf()
-    }
-
-    fn consume(&mut self, amt: usize) {
-        if !self.done_first { self.first.consume(amt) } else { self.second.consume(amt) }
-    }
-
-    fn read_until(&mut self, byte: u8, buf: &mut Vec<u8>) -> Result<usize> {
-        let mut read = 0;
-        if !self.done_first {
-            let n = self.first.read_until(byte, buf)?;
-            read += n;
-
-            match buf.last() {
-                Some(b) if *b == byte && n != 0 => return Ok(read),
-                _ => self.done_first = true,
-            }
-        }
-        read += self.second.read_until(byte, buf)?;
-        Ok(read)
-    }
-
-    // We don't override `read_line` here because an UTF-8 sequence could be
-    // split between the two parts of the chain
-}
-
-impl<T, U> SizeHint for Chain<T, U> {
-    #[inline]
-    fn lower_bound(&self) -> usize {
-        SizeHint::lower_bound(&self.first) + SizeHint::lower_bound(&self.second)
-    }
-
-    #[inline]
-    fn upper_bound(&self) -> Option<usize> {
-        match (SizeHint::upper_bound(&self.first), SizeHint::upper_bound(&self.second)) {
-            (Some(first), Some(second)) => first.checked_add(second),
-            _ => None,
-        }
-    }
-}
-
-/// Reader adapter which limits the bytes read from an underlying reader.
-///
-/// This struct is generally created by calling [`take`] on a reader.
-/// Please see the documentation of [`take`] for more details.
-///
-/// [`take`]: Read::take
-#[stable(feature = "rust1", since = "1.0.0")]
-#[derive(Debug)]
-pub struct Take<T> {
-    inner: T,
-    len: u64,
-    limit: u64,
-}
-
-impl<T> Take<T> {
-    /// Returns the number of bytes that can be read before this instance will
-    /// return EOF.
-    ///
-    /// # Note
-    ///
-    /// This instance may reach `EOF` after reading fewer bytes than indicated by
-    /// this method if the underlying [`Read`] instance reaches EOF.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let f = File::open("foo.txt")?;
-    ///
-    ///     // read at most five bytes
-    ///     let handle = f.take(5);
-    ///
-    ///     println!("limit: {}", handle.limit());
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "rust1", since = "1.0.0")]
-    pub fn limit(&self) -> u64 {
-        self.limit
-    }
-
-    /// Returns the number of bytes read so far.
-    #[unstable(feature = "seek_io_take_position", issue = "97227")]
-    pub fn position(&self) -> u64 {
-        self.len - self.limit
-    }
-
-    /// Sets the number of bytes that can be read before this instance will
-    /// return EOF. This is the same as constructing a new `Take` instance, so
-    /// the amount of bytes read and the previous limit value don't matter when
-    /// calling this method.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let f = File::open("foo.txt")?;
-    ///
-    ///     // read at most five bytes
-    ///     let mut handle = f.take(5);
-    ///     handle.set_limit(10);
-    ///
-    ///     assert_eq!(handle.limit(), 10);
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "take_set_limit", since = "1.27.0")]
-    pub fn set_limit(&mut self, limit: u64) {
-        self.len = limit;
-        self.limit = limit;
-    }
-
-    /// Consumes the `Take`, returning the wrapped reader.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut file = File::open("foo.txt")?;
-    ///
-    ///     let mut buffer = [0; 5];
-    ///     let mut handle = file.take(5);
-    ///     handle.read(&mut buffer)?;
-    ///
-    ///     let file = handle.into_inner();
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "io_take_into_inner", since = "1.15.0")]
-    pub fn into_inner(self) -> T {
-        self.inner
-    }
-
-    /// Gets a reference to the underlying reader.
-    ///
-    /// Care should be taken to avoid modifying the internal I/O state of the
-    /// underlying reader as doing so may corrupt the internal limit of this
-    /// `Take`.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut file = File::open("foo.txt")?;
-    ///
-    ///     let mut buffer = [0; 5];
-    ///     let mut handle = file.take(5);
-    ///     handle.read(&mut buffer)?;
-    ///
-    ///     let file = handle.get_ref();
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "more_io_inner_methods", since = "1.20.0")]
-    pub fn get_ref(&self) -> &T {
-        &self.inner
-    }
-
-    /// Gets a mutable reference to the underlying reader.
-    ///
-    /// Care should be taken to avoid modifying the internal I/O state of the
-    /// underlying reader as doing so may corrupt the internal limit of this
-    /// `Take`.
-    ///
-    /// # Examples
-    ///
-    /// ```no_run
-    /// use std::io;
-    /// use std::io::prelude::*;
-    /// use std::fs::File;
-    ///
-    /// fn main() -> io::Result<()> {
-    ///     let mut file = File::open("foo.txt")?;
-    ///
-    ///     let mut buffer = [0; 5];
-    ///     let mut handle = file.take(5);
-    ///     handle.read(&mut buffer)?;
-    ///
-    ///     let file = handle.get_mut();
-    ///     Ok(())
-    /// }
-    /// ```
-    #[stable(feature = "more_io_inner_methods", since = "1.20.0")]
-    pub fn get_mut(&mut self) -> &mut T {
-        &mut self.inner
-    }
-}
-
-#[stable(feature = "rust1", since = "1.0.0")]
-impl<T: Read> Read for Take<T> {
-    fn read(&mut self, buf: &mut [u8]) -> Result<usize> {
-        // Don't call into inner reader at all at EOF because it may still block
-        if self.limit == 0 {
-            return Ok(0);
-        }
-
-        let max = cmp::min(buf.len() as u64, self.limit) as usize;
-        let n = self.inner.read(&mut buf[..max])?;
-        assert!(n as u64 <= self.limit, "number of read bytes exceeds limit");
-        self.limit -= n as u64;
-        Ok(n)
-    }
-
-    fn read_buf(&mut self, mut buf: BorrowedCursor<'_>) -> Result<()> {
-        // Don't call into inner reader at all at EOF because it may still block
-        if self.limit == 0 {
-            return Ok(());
-        }
-
-        if self.limit < buf.capacity() as u64 {
-            // The condition above guarantees that `self.limit` fits in `usize`.
-            let limit = self.limit as usize;
-
-            let extra_init = cmp::min(limit, buf.init_mut().len());
-
-            // SAFETY: no uninit data is written to ibuf
-            let ibuf = unsafe { &mut buf.as_mut()[..limit] };
-
-            let mut sliced_buf: BorrowedBuf<'_> = ibuf.into();
-
-            // SAFETY: extra_init bytes of ibuf are known to be initialized
-            unsafe {
-                sliced_buf.set_init(extra_init);
-            }
-
-            let mut cursor = sliced_buf.unfilled();
-            let result = self.inner.read_buf(cursor.reborrow());
-
-            let new_init = cursor.init_mut().len();
-            let filled = sliced_buf.len();
-
-            // cursor / sliced_buf / ibuf must drop here
-
-            unsafe {
-                // SAFETY: filled bytes have been filled and therefore initialized
-                buf.advance_unchecked(filled);
-                // SAFETY: new_init bytes of buf's unfilled buffer have been initialized
-                buf.set_init(new_init);
-            }
-
-            self.limit -= filled as u64;
-
-            result
-        } else {
-            let written = buf.written();
-            let result = self.inner.read_buf(buf.reborrow());
-            self.limit -= (buf.written() - written) as u64;
-            result
-        }
-    }
-}
-
-#[stable(feature = "rust1", since = "1.0.0")]
-impl<T: BufRead> BufRead for Take<T> {
-    fn fill_buf(&mut self) -> Result<&[u8]> {
-        // Don't call into inner reader at all at EOF because it may still block
-        if self.limit == 0 {
-            return Ok(&[]);
-        }
-
-        let buf = self.inner.fill_buf()?;
-        let cap = cmp::min(buf.len() as u64, self.limit) as usize;
-        Ok(&buf[..cap])
-    }
-
-    fn consume(&mut self, amt: usize) {
-        // Don't let callers reset the limit by passing an overlarge value
-        let amt = cmp::min(amt as u64, self.limit) as usize;
-        self.limit -= amt as u64;
-        self.inner.consume(amt);
-    }
-}
-
-impl<T> SizeHint for Take<T> {
-    #[inline]
-    fn lower_bound(&self) -> usize {
-        cmp::min(SizeHint::lower_bound(&self.inner) as u64, self.limit) as usize
-    }
-
-    #[inline]
-    fn upper_bound(&self) -> Option<usize> {
-        match SizeHint::upper_bound(&self.inner) {
-            Some(upper_bound) => Some(cmp::min(upper_bound as u64, self.limit) as usize),
-            None => self.limit.try_into().ok(),
-        }
-    }
-}
-
-#[stable(feature = "seek_io_take", since = "1.89.0")]
-impl<T: Seek> Seek for Take<T> {
-    fn seek(&mut self, pos: SeekFrom) -> Result<u64> {
-        let new_position = match pos {
-            SeekFrom::Start(v) => Some(v),
-            SeekFrom::Current(v) => self.position().checked_add_signed(v),
-            SeekFrom::End(v) => self.len.checked_add_signed(v),
-        };
-        let new_position = match new_position {
-            Some(v) if v <= self.len => v,
-            _ => return Err(ErrorKind::InvalidInput.into()),
-        };
-        while new_position != self.position() {
-            if let Some(offset) = new_position.checked_signed_diff(self.position()) {
-                self.inner.seek_relative(offset)?;
-                self.limit = self.limit.wrapping_sub(offset as u64);
-                break;
-            }
-            let offset = if new_position > self.position() { i64::MAX } else { i64::MIN };
-            self.inner.seek_relative(offset)?;
-            self.limit = self.limit.wrapping_sub(offset as u64);
-        }
-        Ok(new_position)
-    }
-
-    fn stream_len(&mut self) -> Result<u64> {
-        Ok(self.len)
-    }
-
-    fn stream_position(&mut self) -> Result<u64> {
-        Ok(self.position())
-    }
-
-    fn seek_relative(&mut self, offset: i64) -> Result<()> {
-        if !self.position().checked_add_signed(offset).is_some_and(|p| p <= self.len) {
-            return Err(ErrorKind::InvalidInput.into());
-        }
-        self.inner.seek_relative(offset)?;
-        self.limit = self.limit.wrapping_sub(offset as u64);
-        Ok(())
-    }
-}
-
-/// An iterator over `u8` values of a reader.
-///
-/// This struct is generally created by calling [`bytes`] on a reader.
-/// Please see the documentation of [`bytes`] for more details.
-///
-/// [`bytes`]: Read::bytes
-#[stable(feature = "rust1", since = "1.0.0")]
-#[derive(Debug)]
-pub struct Bytes<R> {
-    inner: R,
-}
-
-#[stable(feature = "rust1", since = "1.0.0")]
-impl<R: Read> Iterator for Bytes<R> {
-    type Item = Result<u8>;
-
-    // Not `#[inline]`. This function gets inlined even without it, but having
-    // the inline annotation can result in worse code generation. See #116785.
-    fn next(&mut self) -> Option<Result<u8>> {
-        SpecReadByte::spec_read_byte(&mut self.inner)
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        SizeHint::size_hint(&self.inner)
-    }
-}
-
-/// For the specialization of `Bytes::next`.
-trait SpecReadByte {
-    fn spec_read_byte(&mut self) -> Option<Result<u8>>;
-}
-
-impl<R> SpecReadByte for R
-where
-    Self: Read,
-{
-    #[inline]
-    default fn spec_read_byte(&mut self) -> Option<Result<u8>> {
-        inlined_slow_read_byte(self)
-    }
-}
-
-/// Reads a single byte in a slow, generic way. This is used by the default
-/// `spec_read_byte`.
-#[inline]
-fn inlined_slow_read_byte<R: Read>(reader: &mut R) -> Option<Result<u8>> {
-    let mut byte = 0;
-    loop {
-        return match reader.read(slice::from_mut(&mut byte)) {
-            Ok(0) => None,
-            Ok(..) => Some(Ok(byte)),
-            Err(ref e) if e.is_interrupted() => continue,
-            Err(e) => Some(Err(e)),
-        };
-    }
-}
-
-// Used by `BufReader::spec_read_byte`, for which the `inline(never)` is
-// important.
-#[inline(never)]
-fn uninlined_slow_read_byte<R: Read>(reader: &mut R) -> Option<Result<u8>> {
-    inlined_slow_read_byte(reader)
-}
-
-trait SizeHint {
-    fn lower_bound(&self) -> usize;
-
-    fn upper_bound(&self) -> Option<usize>;
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (self.lower_bound(), self.upper_bound())
-    }
-}
-
-impl<T: ?Sized> SizeHint for T {
-    #[inline]
-    default fn lower_bound(&self) -> usize {
-        0
-    }
-
-    #[inline]
-    default fn upper_bound(&self) -> Option<usize> {
-        None
-    }
-}
-
-impl<T> SizeHint for &mut T {
-    #[inline]
-    fn lower_bound(&self) -> usize {
-        SizeHint::lower_bound(*self)
-    }
-
-    #[inline]
-    fn upper_bound(&self) -> Option<usize> {
-        SizeHint::upper_bound(*self)
-    }
-}
-
-impl<T> SizeHint for Box<T> {
-    #[inline]
-    fn lower_bound(&self) -> usize {
-        SizeHint::lower_bound(&**self)
-    }
-
-    #[inline]
-    fn upper_bound(&self) -> Option<usize> {
-        SizeHint::upper_bound(&**self)
-    }
-}
-
-impl SizeHint for &[u8] {
-    #[inline]
-    fn lower_bound(&self) -> usize {
-        self.len()
-    }
-
-    #[inline]
-    fn upper_bound(&self) -> Option<usize> {
-        Some(self.len())
-    }
-}
-
-/// An iterator over the contents of an instance of `BufRead` split on a
-/// particular byte.
-///
-/// This struct is generally created by calling [`split`] on a `BufRead`.
-/// Please see the documentation of [`split`] for more details.
-///
-/// [`split`]: BufRead::split
-#[stable(feature = "rust1", since = "1.0.0")]
-#[derive(Debug)]
-#[cfg_attr(not(test), rustc_diagnostic_item = "IoSplit")]
-pub struct Split<B> {
-    buf: B,
-    delim: u8,
-}
-
-#[stable(feature = "rust1", since = "1.0.0")]
-impl<B: BufRead> Iterator for Split<B> {
-    type Item = Result<Vec<u8>>;
-
-    fn next(&mut self) -> Option<Result<Vec<u8>>> {
-        let mut buf = Vec::new();
-        match self.buf.read_until(self.delim, &mut buf) {
-            Ok(0) => None,
-            Ok(_n) => {
-                if buf[buf.len() - 1] == self.delim {
-                    buf.pop();
-                }
-                Some(Ok(buf))
-            }
-            Err(e) => Some(Err(e)),
-        }
-    }
-}
-
-/// An iterator over the lines of an instance of `BufRead`.
-///
-/// This struct is generally created by calling [`lines`] on a `BufRead`.
-/// Please see the documentation of [`lines`] for more details.
-///
-/// [`lines`]: BufRead::lines
-#[stable(feature = "rust1", since = "1.0.0")]
-#[derive(Debug)]
-#[cfg_attr(not(test), rustc_diagnostic_item = "IoLines")]
-pub struct Lines<B> {
-    buf: B,
-}
-
-#[stable(feature = "rust1", since = "1.0.0")]
-impl<B: BufRead> Iterator for Lines<B> {
-    type Item = Result<String>;
-
-    fn next(&mut self) -> Option<Result<String>> {
-        let mut buf = String::new();
-        match self.buf.read_line(&mut buf) {
-            Ok(0) => None,
-            Ok(_n) => {
-                if buf.ends_with('\n') {
-                    buf.pop();
-                    if buf.ends_with('\r') {
-                        buf.pop();
-                    }
-                }
-                Some(Ok(buf))
-            }
-            Err(e) => Some(Err(e)),
-        }
-    }
-}
