+++ library/std/src/lib.rs
@@ -373,19 +373,25 @@
 // tidy-alphabetical-start
 #![feature(alloc_io)]
 #![feature(allocator_api)]
+#![feature(can_vector)]
 #![feature(clone_from_ref)]
 #![feature(get_mut_unchecked)]
 #![feature(io_error_inprogress)]
 #![feature(io_error_internals)]
 #![feature(io_error_more)]
 #![feature(io_error_uncategorized)]
+#![feature(io_internals)]
 #![feature(io_slice_as_bytes)]
 #![feature(map_try_insert)]
+#![feature(raw_os_error_ty)]
+#![feature(read_buf)]
+#![feature(seek_stream_len)]
 #![feature(slice_concat_trait)]
 #![feature(thin_box)]
 #![feature(try_reserve_kind)]
 #![feature(try_with_capacity)]
 #![feature(unique_rc_arc)]
+#![feature(write_all_vectored)]
 #![feature(wtf8_internals)]
 // tidy-alphabetical-end
 //
