+++ library/std/src/sys/io/mod.rs
@@ -44,7 +44,3 @@ pub use error::set_errno;
 pub use error::{decode_error_kind, errno, error_string, is_interrupted};
 pub use is_terminal::is_terminal;
 pub use kernel_copy::{CopyState, kernel_copy};
-
-// Bare metal platforms usually have very small amounts of RAM
-// (in the order of hundreds of KB)
-pub const DEFAULT_BUF_SIZE: usize = if cfg!(target_os = "espidf") { 512 } else { 8 * 1024 };
