+++ library/std/src/sys/stdio/motor.rs
@@ -2,7 +2,7 @@ use crate::os::fd::{AsFd, AsRawFd, BorrowedFd, FromRawFd, IntoRawFd, OwnedFd, Ra
 use crate::sys::{AsInner, FromInner, IntoInner, map_motor_error};
 use crate::{io, process, sys};
 
-pub const STDIN_BUF_SIZE: usize = crate::sys::io::DEFAULT_BUF_SIZE;
+pub const STDIN_BUF_SIZE: usize = alloc::io::DEFAULT_BUF_SIZE;
 
 pub struct Stdin {}
